+++ stdlib/public/core/DictionaryBridging.swift
@@ -1,0 +1,758 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2014 - 2018 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+#if _runtime(_ObjC)
+
+import SwiftShims
+
+/// Equivalent to `NSDictionary.allKeys`, but does not leave objects on the
+/// autorelease pool.
+@inlinable
+internal func _stdlib_NSDictionary_allKeys(
+  _ nsd: _NSDictionary
+) -> _HeapBuffer<Int, AnyObject> {
+  let count = nsd.count
+  let storage = _HeapBuffer<Int, AnyObject>(
+    _HeapBufferStorage<Int, AnyObject>.self, count, count)
+
+  nsd.getObjects(nil, andKeys: storage.baseAddress, count: count)
+  return storage
+}
+
+extension _NativeDictionary { // Bridging
+  @usableFromInline
+  internal func bridged() -> _NSDictionary {
+    // We can zero-cost bridge if our keys are verbatim
+    // or if we're the empty singleton.
+
+    // Temporary var for SOME type safety before a cast.
+    let nsDictionary: _NSDictionaryCore
+
+    if _storage === _RawDictionaryStorage.empty || count == 0 {
+      nsDictionary = _RawDictionaryStorage.empty
+    } else if _isBridgedVerbatimToObjectiveC(Key.self),
+      _isBridgedVerbatimToObjectiveC(Value.self) {
+      nsDictionary = unsafeDowncast(
+        _storage,
+        to: _DictionaryStorage<Key, Value>.self)
+    } else {
+      nsDictionary = _SwiftDeferredNSDictionary(self)
+    }
+
+    // Cast from "minimal NSDictionary" to "NSDictionary"
+    // Note that if you actually ask Swift for this cast, it will fail.
+    // Never trust a shadow protocol!
+    return unsafeBitCast(nsDictionary, to: _NSDictionary.self)
+  }
+}
+
+/// An NSEnumerator that works with any _NativeDictionary of
+/// verbatim bridgeable elements. Used by the various NSDictionary impls.
+final internal class _SwiftDictionaryNSEnumerator<Key: Hashable, Value>
+  : _SwiftNativeNSEnumerator, _NSEnumerator {
+
+  @nonobjc internal var base: _NativeDictionary<Key, Value>
+  @nonobjc internal var bridgedKeys: _BridgingHashBuffer?
+  @nonobjc internal var nextIndex: _NativeDictionary<Key, Value>.Index
+  @nonobjc internal var endIndex: _NativeDictionary<Key, Value>.Index
+
+  @objc
+  internal override required init() {
+    _sanityCheckFailure("don't call this designated initializer")
+  }
+
+  internal init(_ base: _NativeDictionary<Key, Value>) {
+    _sanityCheck(_isBridgedVerbatimToObjectiveC(Key.self))
+    self.base = base
+    self.bridgedKeys = nil
+    self.nextIndex = base.startIndex
+    self.endIndex = base.endIndex
+  }
+
+  @nonobjc
+  internal init(_ deferred: _SwiftDeferredNSDictionary<Key, Value>) {
+    _sanityCheck(!_isBridgedVerbatimToObjectiveC(Key.self))
+    self.base = deferred.native
+    self.bridgedKeys = deferred.bridgeKeys()
+    self.nextIndex = base.startIndex
+    self.endIndex = base.endIndex
+  }
+
+  private func bridgedKey(at index: _HashTable.Index) -> AnyObject {
+    _sanityCheck(base.hashTable.isOccupied(index))
+    if let bridgedKeys = self.bridgedKeys {
+      return bridgedKeys[index]
+    }
+    return _bridgeAnythingToObjectiveC(base.uncheckedKey(at: index))
+  }
+
+  @objc
+  internal func nextObject() -> AnyObject? {
+    if nextIndex == endIndex {
+      return nil
+    }
+    let index = nextIndex
+    nextIndex = base.index(after: nextIndex)
+    return self.bridgedKey(at: index)
+  }
+
+  @objc(countByEnumeratingWithState:objects:count:)
+  internal func countByEnumerating(
+    with state: UnsafeMutablePointer<_SwiftNSFastEnumerationState>,
+    objects: UnsafeMutablePointer<AnyObject>,
+    count: Int
+  ) -> Int {
+    var theState = state.pointee
+    if theState.state == 0 {
+      theState.state = 1 // Arbitrary non-zero value.
+      theState.itemsPtr = AutoreleasingUnsafeMutablePointer(objects)
+      theState.mutationsPtr = _fastEnumerationStorageMutationsPtr
+    }
+
+    if nextIndex == endIndex {
+      state.pointee = theState
+      return 0
+    }
+
+    // Return only a single element so that code can start iterating via fast
+    // enumeration, terminate it, and continue via NSEnumerator.
+    let unmanagedObjects = _UnmanagedAnyObjectArray(objects)
+    unmanagedObjects[0] = self.bridgedKey(at: nextIndex)
+    nextIndex = base.index(after: nextIndex)
+    state.pointee = theState
+    return 1
+  }
+}
+
+/// This class exists for Objective-C bridging. It holds a reference to a
+/// _NativeDictionary, and can be upcast to NSSelf when bridging is
+/// necessary.  This is the fallback implementation for situations where
+/// toll-free bridging isn't possible. On first access, a _NativeDictionary
+/// of AnyObject will be constructed containing all the bridged elements.
+final internal class _SwiftDeferredNSDictionary<Key: Hashable, Value>
+  : _SwiftNativeNSDictionary, _NSDictionaryCore {
+
+  // This stored property must be stored at offset zero.  We perform atomic
+  // operations on it.
+  //
+  // Do not access this property directly.
+  @nonobjc
+  private var _bridgedKeys_DoNotUse: AnyObject?
+
+  // This stored property must be stored at offset one.  We perform atomic
+  // operations on it.
+  //
+  // Do not access this property directly.
+  @nonobjc
+  private var _bridgedValues_DoNotUse: AnyObject?
+
+  /// The unbridged elements.
+  internal var native: _NativeDictionary<Key, Value>
+
+  internal init(_ native: _NativeDictionary<Key, Value>) {
+    _sanityCheck(native.count > 0)
+    _sanityCheck(!_isBridgedVerbatimToObjectiveC(Key.self) ||
+      !_isBridgedVerbatimToObjectiveC(Value.self))
+    self.native = native
+    super.init()
+  }
+
+  @objc
+  internal required init(
+    objects: UnsafePointer<AnyObject?>,
+    forKeys: UnsafeRawPointer,
+    count: Int
+  ) {
+    _sanityCheckFailure("don't call this designated initializer")
+  }
+
+  @nonobjc
+  private var _bridgedKeysPtr: UnsafeMutablePointer<AnyObject?> {
+    return _getUnsafePointerToStoredProperties(self)
+      .assumingMemoryBound(to: Optional<AnyObject>.self)
+  }
+
+  @nonobjc
+  private var _bridgedValuesPtr: UnsafeMutablePointer<AnyObject?> {
+    return _bridgedKeysPtr + 1
+  }
+
+  /// The buffer for bridged keys, if present.
+  @nonobjc
+  private var _bridgedKeys: _BridgingHashBuffer? {
+    guard let ref = _stdlib_atomicLoadARCRef(object: _bridgedKeysPtr) else {
+      return nil
+    }
+    return unsafeDowncast(ref, to: _BridgingHashBuffer.self)
+  }
+
+  /// The buffer for bridged values, if present.
+  @nonobjc
+  private var _bridgedValues: _BridgingHashBuffer? {
+    guard let ref = _stdlib_atomicLoadARCRef(object: _bridgedValuesPtr) else {
+      return nil
+    }
+    return unsafeDowncast(ref, to: _BridgingHashBuffer.self)
+  }
+
+  /// Attach a buffer for bridged Dictionary keys.
+  @nonobjc
+  private func _initializeBridgedKeys(_ storage: _BridgingHashBuffer) {
+    _stdlib_atomicInitializeARCRef(object: _bridgedKeysPtr, desired: storage)
+  }
+
+  /// Attach a buffer for bridged Dictionary values.
+  @nonobjc
+  private func _initializeBridgedValues(_ storage: _BridgingHashBuffer) {
+    _stdlib_atomicInitializeARCRef(object: _bridgedValuesPtr, desired: storage)
+  }
+
+  @nonobjc
+  internal func bridgeKeys() -> _BridgingHashBuffer? {
+    if _isBridgedVerbatimToObjectiveC(Key.self) { return nil }
+    if let bridgedKeys = _bridgedKeys { return bridgedKeys }
+
+    // Allocate and initialize heap storage for bridged keys.
+    let bridged = _BridgingHashBuffer.allocate(
+      owner: native._storage,
+      hashTable: native.hashTable)
+    for index in native.hashTable {
+      let object = _bridgeAnythingToObjectiveC(native.uncheckedKey(at: index))
+      bridged.initialize(at: index, to: object)
+    }
+
+    // Atomically put the bridged keys in place.
+    _initializeBridgedKeys(bridged)
+    return _bridgedKeys!
+  }
+
+  @nonobjc
+  internal func bridgeValues() -> _BridgingHashBuffer? {
+    if _isBridgedVerbatimToObjectiveC(Value.self) { return nil }
+    if let bridgedValues = _bridgedValues { return bridgedValues }
+
+    // Allocate and initialize heap storage for bridged values.
+    let bridged = _BridgingHashBuffer.allocate(
+      owner: native._storage,
+      hashTable: native.hashTable)
+    for index in native.hashTable {
+      let object = _bridgeAnythingToObjectiveC(native.uncheckedValue(at: index))
+      bridged.initialize(at: index, to: object)
+    }
+
+    // Atomically put the bridged values in place.
+    _initializeBridgedValues(bridged)
+    return _bridgedValues!
+  }
+
+  @usableFromInline
+  internal typealias Index = _HashTable.Index
+
+  @objc(copyWithZone:)
+  internal func copy(with zone: _SwiftNSZone?) -> AnyObject {
+    // Instances of this class should be visible outside of standard library as
+    // having `NSDictionary` type, which is immutable.
+    return self
+  }
+
+  @objc(objectForKey:)
+  internal func object(forKey aKey: AnyObject) -> AnyObject? {
+    guard let nativeKey = _conditionallyBridgeFromObjectiveC(aKey, Key.self)
+    else { return nil }
+
+    let (index, found) = native.find(nativeKey)
+    guard found else { return nil }
+    if let bridgedValues = bridgeValues() {
+      return bridgedValues[index]
+    }
+    return _bridgeAnythingToObjectiveC(native.uncheckedValue(at: index))
+  }
+
+  @inline(__always)
+  private func _key(
+    at index: Index,
+    bridgedKeys: _BridgingHashBuffer?
+  ) -> AnyObject {
+    if let bridgedKeys = bridgedKeys {
+      return bridgedKeys[index]
+    }
+    return _bridgeAnythingToObjectiveC(native.uncheckedKey(at: index))
+  }
+
+  @inline(__always)
+  private func _value(
+    at index: Index,
+    bridgedValues: _BridgingHashBuffer?
+  ) -> AnyObject {
+    if let bridgedValues = bridgedValues {
+      return bridgedValues[index]
+    }
+    return _bridgeAnythingToObjectiveC(native.uncheckedValue(at: index))
+  }
+
+  @objc
+  internal func keyEnumerator() -> _NSEnumerator {
+    if _isBridgedVerbatimToObjectiveC(Key.self) {
+      return _SwiftDictionaryNSEnumerator<Key, Value>(native)
+    }
+    return _SwiftDictionaryNSEnumerator<Key, Value>(self)
+  }
+
+  @objc(getObjects:andKeys:count:)
+  internal func getObjects(
+    _ objects: UnsafeMutablePointer<AnyObject>?,
+    andKeys keys: UnsafeMutablePointer<AnyObject>?,
+    count: Int
+  ) {
+    _precondition(count >= 0, "Invalid count")
+    guard count > 0 else { return }
+    let bridgedKeys = bridgeKeys()
+    let bridgedValues = bridgeValues()
+    var i = 0 // Current position in the output buffers
+    let bucketCount = native._storage._bucketCount
+
+    defer { _fixLifetime(self) }
+
+    switch (_UnmanagedAnyObjectArray(keys), _UnmanagedAnyObjectArray(objects)) {
+    case (let unmanagedKeys?, let unmanagedObjects?):
+      for index in native.hashTable {
+        unmanagedKeys[i] = _key(at: index, bridgedKeys: bridgedKeys)
+        unmanagedObjects[i] = _value(at: index, bridgedValues: bridgedValues)
+        i += 1
+        guard i < count else { break }
+      }
+    case (let unmanagedKeys?, nil):
+      for index in native.hashTable {
+        unmanagedKeys[i] = _key(at: index, bridgedKeys: bridgedKeys)
+        i += 1
+        guard i < count else { break }
+      }
+    case (nil, let unmanagedObjects?):
+      for index in native.hashTable {
+        unmanagedObjects[i] = _value(at: index, bridgedValues: bridgedValues)
+        i += 1
+        guard i < count else { break }
+      }
+    case (nil, nil):
+      // Do nothing
+      break
+    }
+  }
+
+  @objc(enumerateKeysAndObjectsWithOptions:usingBlock:)
+  internal func enumerateKeysAndObjects(
+    options: Int,
+    using block: @convention(block) (
+      Unmanaged<AnyObject>,
+      Unmanaged<AnyObject>,
+      UnsafeMutablePointer<UInt8>
+    ) -> Void) {
+    let bridgedKeys = bridgeKeys()
+    let bridgedValues = bridgeValues()
+
+    defer { _fixLifetime(self) }
+
+    var stop: UInt8 = 0
+    for index in native.hashTable {
+      let key = _key(at: index, bridgedKeys: bridgedKeys)
+      let value = _value(at: index, bridgedValues: bridgedValues)
+      block(
+        Unmanaged.passUnretained(key),
+        Unmanaged.passUnretained(value),
+        &stop)
+      if stop != 0 { return }
+    }
+  }
+
+  @objc
+  internal var count: Int {
+    return native.count
+  }
+
+  @objc(countByEnumeratingWithState:objects:count:)
+  internal func countByEnumerating(
+    with state: UnsafeMutablePointer<_SwiftNSFastEnumerationState>,
+    objects: UnsafeMutablePointer<AnyObject>?,
+    count: Int
+  ) -> Int {
+    var theState = state.pointee
+    if theState.state == 0 {
+      theState.state = 1 // Arbitrary non-zero value.
+      theState.itemsPtr = AutoreleasingUnsafeMutablePointer(objects)
+      theState.mutationsPtr = _fastEnumerationStorageMutationsPtr
+      theState.extra.0 = CUnsignedLong(native.startIndex.bucket)
+    }
+
+    // Test 'objects' rather than 'count' because (a) this is very rare anyway,
+    // and (b) the optimizer should then be able to optimize away the
+    // unwrapping check below.
+    if _slowPath(objects == nil) {
+      return 0
+    }
+
+    let unmanagedObjects = _UnmanagedAnyObjectArray(objects!)
+    var index = _HashTable.Index(bucket: Int(theState.extra.0))
+    let endIndex = native.endIndex
+    _precondition(index == endIndex || native.hashTable.isOccupied(index))
+    var stored = 0
+
+    // Only need to bridge once, so we can hoist it out of the loop.
+    let bridgedKeys = bridgeKeys()
+    for i in 0..<count {
+      if index == endIndex { break }
+
+      unmanagedObjects[i] = _key(at: index, bridgedKeys: bridgedKeys)
+      stored += 1
+      index = native.index(after: index)
+    }
+    theState.extra.0 = CUnsignedLong(index.bucket)
+    state.pointee = theState
+    return stored
+  }
+}
+
+@usableFromInline
+@_fixed_layout
+internal struct _CocoaDictionary {
+  @usableFromInline
+  internal let object: _NSDictionary
+
+  @inlinable
+  internal init(_ object: _NSDictionary) {
+    self.object = object
+  }
+}
+
+extension _CocoaDictionary: Equatable {
+  @usableFromInline
+  internal static func ==(
+    lhs: _CocoaDictionary,
+    rhs: _CocoaDictionary
+  ) -> Bool {
+    return _stdlib_NSObject_isEqual(lhs.object, rhs.object)
+  }
+}
+
+extension _CocoaDictionary: _DictionaryBuffer {
+  @usableFromInline
+  internal typealias Key = AnyObject
+  @usableFromInline
+  internal typealias Value = AnyObject
+
+  @inlinable
+  internal var startIndex: Index {
+    return Index(self, startIndex: ())
+  }
+
+  @inlinable
+  internal var endIndex: Index {
+    return Index(self, endIndex: ())
+  }
+
+  @inlinable
+  internal func index(after i: Index) -> Index {
+    var i = i
+    formIndex(after: &i)
+    return i
+  }
+
+  @usableFromInline
+  @_effects(releasenone)
+  internal func formIndex(after i: inout Index) {
+    _precondition(i.base.object === self.object, "Invalid index")
+    _precondition(i.currentKeyIndex < i.allKeys.value,
+      "Cannot increment endIndex")
+    i.currentKeyIndex += 1
+  }
+
+  @usableFromInline
+  internal func index(forKey key: Key) -> Index? {
+    // Fast path that does not involve creating an array of all keys.  In case
+    // the key is present, this lookup is a penalty for the slow path, but the
+    // potential savings are significant: we could skip a memory allocation and
+    // a linear search.
+    if lookup(key) == nil {
+      return nil
+    }
+
+    let allKeys = _stdlib_NSDictionary_allKeys(object)
+    var keyIndex = -1
+    for i in 0..<allKeys.value {
+      if _stdlib_NSObject_isEqual(key, allKeys[i]) {
+        keyIndex = i
+        break
+      }
+    }
+    _sanityCheck(keyIndex >= 0,
+        "Key was found in fast path, but not found later?")
+    return Index(self, allKeys, keyIndex)
+  }
+
+  @inlinable
+  internal var count: Int {
+    return object.count
+  }
+
+  @inlinable
+  @inline(__always)
+  internal func contains(_ key: Key) -> Bool {
+    return object.object(forKey: key) != nil
+  }
+
+  @inlinable
+  @inline(__always)
+  internal func lookup(_ key: Key) -> Value? {
+    return object.object(forKey: key)
+  }
+
+  @inlinable
+  @inline(__always)
+  internal func lookup(_ index: Index) -> (key: Key, value: Value) {
+    _precondition(index.base.object === self.object, "Invalid index")
+    let key: Key = index.allKeys[index.currentKeyIndex]
+    let value: Value = index.base.object.object(forKey: key)!
+    return (key, value)
+  }
+
+  @inlinable
+  @inline(__always)
+  func key(at index: Index) -> Key {
+    _precondition(index.base.object === self.object, "Invalid index")
+    return index.allKeys[index.currentKeyIndex]
+  }
+
+  @inlinable
+  @inline(__always)
+  func value(at index: Index) -> Value {
+    _precondition(index.base.object === self.object, "Invalid index")
+    let key = index.allKeys[index.currentKeyIndex]
+    return index.base.object.object(forKey: key)!
+  }
+}
+
+extension _CocoaDictionary {
+  @inlinable
+  internal func mapValues<Key: Hashable, Value, T>(
+    _ transform: (Value) throws -> T
+  ) rethrows -> _NativeDictionary<Key, T> {
+    var result = _NativeDictionary<Key, T>(capacity: self.count)
+    for (cocoaKey, cocoaValue) in self {
+      let key = _forceBridgeFromObjectiveC(cocoaKey, Key.self)
+      let value = _forceBridgeFromObjectiveC(cocoaValue, Value.self)
+      try result.insertNew(key: key, value: transform(value))
+    }
+    return result
+  }
+}
+
+extension _CocoaDictionary {
+  @_fixed_layout // FIXME(sil-serialize-all)
+  @usableFromInline
+  internal struct Index {
+    // Assumption: we rely on NSDictionary.getObjects when being
+    // repeatedly called on the same NSDictionary, returning items in the same
+    // order every time.
+    // Similarly, the same assumption holds for NSSet.allObjects.
+
+    /// A reference to the NSDictionary, which owns members in `allObjects`,
+    /// or `allKeys`, for NSSet and NSDictionary respectively.
+    @usableFromInline // FIXME(sil-serialize-all)
+    internal let base: _CocoaDictionary
+    // FIXME: swift-3-indexing-model: try to remove the cocoa reference, but
+    // make sure that we have a safety check for accessing `allKeys`.  Maybe
+    // move both into the dictionary/set itself.
+
+    /// An unowned array of keys.
+    @usableFromInline // FIXME(sil-serialize-all)
+    internal var allKeys: _HeapBuffer<Int, AnyObject>
+
+    /// Index into `allKeys`
+    @usableFromInline // FIXME(sil-serialize-all)
+    internal var currentKeyIndex: Int
+
+    @inlinable // FIXME(sil-serialize-all)
+    internal init(_ base: _CocoaDictionary, startIndex: ()) {
+      self.base = base
+      self.allKeys = _stdlib_NSDictionary_allKeys(base.object)
+      self.currentKeyIndex = 0
+    }
+
+    @inlinable // FIXME(sil-serialize-all)
+    internal init(_ base: _CocoaDictionary, endIndex: ()) {
+      self.base = base
+      self.allKeys = _stdlib_NSDictionary_allKeys(base.object)
+      self.currentKeyIndex = allKeys.value
+    }
+
+    @inlinable // FIXME(sil-serialize-all)
+    internal init(
+      _ base: _CocoaDictionary,
+      _ allKeys: _HeapBuffer<Int, AnyObject>,
+      _ currentKeyIndex: Int
+    ) {
+      self.base = base
+      self.allKeys = allKeys
+      self.currentKeyIndex = currentKeyIndex
+    }
+  }
+}
+
+extension _CocoaDictionary.Index: Equatable {
+  @inlinable
+  internal static func == (
+    lhs: _CocoaDictionary.Index,
+    rhs: _CocoaDictionary.Index
+  ) -> Bool {
+    _precondition(lhs.base.object === rhs.base.object,
+      "Comparing indexes from different dictionaries")
+    return lhs.currentKeyIndex == rhs.currentKeyIndex
+  }
+}
+
+extension _CocoaDictionary.Index: Comparable {
+  @inlinable
+  internal static func < (
+    lhs: _CocoaDictionary.Index,
+    rhs: _CocoaDictionary.Index
+  ) -> Bool {
+    _precondition(lhs.base.object === rhs.base.object,
+      "Comparing indexes from different dictionaries")
+    return lhs.currentKeyIndex < rhs.currentKeyIndex
+  }
+}
+
+extension _CocoaDictionary: Sequence {
+  @usableFromInline
+  final internal class Iterator {
+    // Cocoa Dictionary iterator has to be a class, otherwise we cannot
+    // guarantee that the fast enumeration struct is pinned to a certain memory
+    // location.
+
+    // This stored property should be stored at offset zero.  There's code below
+    // relying on this.
+    internal var _fastEnumerationState: _SwiftNSFastEnumerationState =
+      _makeSwiftNSFastEnumerationState()
+
+    // This stored property should be stored right after
+    // `_fastEnumerationState`.  There's code below relying on this.
+    internal var _fastEnumerationStackBuf = _CocoaFastEnumerationStackBuf()
+
+    internal let base: _CocoaDictionary
+
+    internal var _fastEnumerationStatePtr:
+      UnsafeMutablePointer<_SwiftNSFastEnumerationState> {
+      return _getUnsafePointerToStoredProperties(self).assumingMemoryBound(
+        to: _SwiftNSFastEnumerationState.self)
+    }
+
+    internal var _fastEnumerationStackBufPtr:
+      UnsafeMutablePointer<_CocoaFastEnumerationStackBuf> {
+      return UnsafeMutableRawPointer(_fastEnumerationStatePtr + 1)
+      .assumingMemoryBound(to: _CocoaFastEnumerationStackBuf.self)
+    }
+
+    // These members have to be word-sized integers, they cannot be limited to
+    // Int8 just because our storage holds 16 elements: fast enumeration is
+    // allowed to return inner pointers to the container, which can be much
+    // larger.
+    internal var itemIndex: Int = 0
+    internal var itemCount: Int = 0
+
+    internal init(_ base: _CocoaDictionary) {
+      self.base = base
+    }
+  }
+
+  @usableFromInline
+  @_effects(releasenone)
+  internal func makeIterator() -> Iterator {
+    return Iterator(self)
+  }
+}
+
+extension _CocoaDictionary.Iterator: IteratorProtocol {
+  @usableFromInline
+  internal typealias Element = (key: AnyObject, value: AnyObject)
+
+  @usableFromInline
+  internal func next() -> Element? {
+    if itemIndex < 0 {
+      return nil
+    }
+    let base = self.base
+    if itemIndex == itemCount {
+      let stackBufCount = _fastEnumerationStackBuf.count
+      // We can't use `withUnsafeMutablePointer` here to get pointers to
+      // properties, because doing so might introduce a writeback storage, but
+      // fast enumeration relies on the pointer identity of the enumeration
+      // state struct.
+      itemCount = base.object.countByEnumerating(
+        with: _fastEnumerationStatePtr,
+        objects: UnsafeMutableRawPointer(_fastEnumerationStackBufPtr)
+          .assumingMemoryBound(to: AnyObject.self),
+        count: stackBufCount)
+      if itemCount == 0 {
+        itemIndex = -1
+        return nil
+      }
+      itemIndex = 0
+    }
+    let itemsPtrUP =
+    UnsafeMutableRawPointer(_fastEnumerationState.itemsPtr!)
+      .assumingMemoryBound(to: AnyObject.self)
+    let itemsPtr = _UnmanagedAnyObjectArray(itemsPtrUP)
+    let key: AnyObject = itemsPtr[itemIndex]
+    itemIndex += 1
+    let value: AnyObject = base.object.object(forKey: key)!
+    return (key, value)
+  }
+}
+
+//===--- Bridging ---------------------------------------------------------===//
+
+extension Dictionary {
+  @inlinable
+  public func _bridgeToObjectiveCImpl() -> _NSDictionaryCore {
+    switch _variant {
+    case .native(let nativeDictionary):
+      return nativeDictionary.bridged()
+    case .cocoa(let cocoaDictionary):
+      return cocoaDictionary.object
+    }
+  }
+
+  /// Returns the native Dictionary hidden inside this NSDictionary;
+  /// returns nil otherwise.
+  public static func _bridgeFromObjectiveCAdoptingNativeStorageOf(
+    _ s: AnyObject
+  ) -> Dictionary<Key, Value>? {
+
+    // Try all three NSDictionary impls that we currently provide.
+
+    if let deferred = s as? _SwiftDeferredNSDictionary<Key, Value> {
+      return Dictionary(_native: deferred.native)
+    }
+
+    if let nativeStorage = s as? _DictionaryStorage<Key, Value> {
+      return Dictionary(_native: _NativeDictionary(nativeStorage))
+    }
+
+    if s === _RawDictionaryStorage.empty {
+      return Dictionary()
+    }
+
+    // FIXME: what if `s` is native storage, but for different key/value type?
+    return nil
+  }
+}
+
+#endif // _runtime(_ObjC)
