+++ stdlib/public/core/Set.swift
@@ -10,8 +10,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-import SwiftShims
-
 /// An unordered collection of unique elements.
 ///
 /// You use a set instead of an array when you need to test efficiently for
@@ -1062,110 +1060,6 @@ extension Set: CustomStringConvertible, CustomDebugStringConvertible {
   }
 }
 
-#if _runtime(_ObjC)
-@_silgen_name("swift_stdlib_CFSetGetValues")
-@usableFromInline
-internal
-func _stdlib_CFSetGetValues(_ nss: _NSSet, _: UnsafeMutablePointer<AnyObject>)
-
-/// Equivalent to `NSSet.allObjects`, but does not leave objects on the
-/// autorelease pool.
-@inlinable
-internal func _stdlib_NSSet_allObjects(
-  _ nss: _NSSet
-) -> _HeapBuffer<Int, AnyObject> {
-  let count = nss.count
-  let storage = _HeapBuffer<Int, AnyObject>(
-    _HeapBufferStorage<Int, AnyObject>.self, count, count)
-  _stdlib_CFSetGetValues(nss, storage.baseAddress)
-  return storage
-}
-#endif
-
-//===--- Compiler conversion/casting entry points for Set<Element> --------===//
-
-/// Perform a non-bridged upcast that always succeeds.
-///
-/// - Precondition: `BaseValue` is a base class or base `@objc`
-///   protocol (such as `AnyObject`) of `DerivedValue`.
-@inlinable
-public func _setUpCast<DerivedValue, BaseValue>(_ source: Set<DerivedValue>)
-  -> Set<BaseValue> {
-  var builder = _SetBuilder<BaseValue>(count: source.count)
-  for x in source {
-    builder.add(member: x as! BaseValue)
-  }
-  return builder.take()
-}
-
-/// Called by the casting machinery.
-@_silgen_name("_swift_setDownCastIndirect")
-internal func _setDownCastIndirect<SourceValue, TargetValue>(
-  _ source: UnsafePointer<Set<SourceValue>>,
-  _ target: UnsafeMutablePointer<Set<TargetValue>>) {
-  target.initialize(to: _setDownCast(source.pointee))
-}
-
-/// Implements a forced downcast.  This operation should have O(1) complexity.
-///
-/// The cast can fail if bridging fails.  The actual checks and bridging can be
-/// deferred.
-///
-/// - Precondition: `DerivedValue` is a subtype of `BaseValue` and both
-///   are reference types.
-@inlinable
-public func _setDownCast<BaseValue, DerivedValue>(_ source: Set<BaseValue>)
-  -> Set<DerivedValue> {
-
-#if _runtime(_ObjC)
-  if _isClassOrObjCExistential(BaseValue.self)
-  && _isClassOrObjCExistential(DerivedValue.self) {
-    switch source._variant {
-    case .native(let nativeSet):
-      return Set(_immutableCocoaSet: nativeSet.bridged())
-    case .cocoa(let cocoaSet):
-      return Set(_immutableCocoaSet: cocoaSet.object)
-    }
-  }
-#endif
-  return _setDownCastConditional(source)!
-}
-
-/// Called by the casting machinery.
-@_silgen_name("_swift_setDownCastConditionalIndirect")
-internal func _setDownCastConditionalIndirect<SourceValue, TargetValue>(
-  _ source: UnsafePointer<Set<SourceValue>>,
-  _ target: UnsafeMutablePointer<Set<TargetValue>>
-) -> Bool {
-  if let result: Set<TargetValue> = _setDownCastConditional(source.pointee) {
-    target.initialize(to: result)
-    return true
-  }
-  return false
-}
-
-/// Implements a conditional downcast.
-///
-/// If the cast fails, the function returns `nil`.  All checks should be
-/// performed eagerly.
-///
-/// - Precondition: `DerivedValue` is a subtype of `BaseValue` and both
-///   are reference types.
-@inlinable
-public func _setDownCastConditional<BaseValue, DerivedValue>(
-  _ source: Set<BaseValue>
-) -> Set<DerivedValue>? {
-  var result = Set<DerivedValue>(minimumCapacity: source.count)
-  for member in source {
-    if let derivedMember = member as? DerivedValue {
-      result.insert(derivedMember)
-      continue
-    }
-    return nil
-  }
-  return result
-}
-
 extension Set {
   /// Removes the elements of the given set from this set.
   ///
@@ -1373,1532 +1267,6 @@ extension Set {
   }
 }
 
-/// This protocol is only used for compile-time checks that
-/// every buffer type implements all required operations.
-internal protocol _SetBuffer {
-  associatedtype Element
-  associatedtype Index
-
-  var startIndex: Index { get }
-  var endIndex: Index { get }
-  func index(after i: Index) -> Index
-  func index(for element: Element) -> Index?
-  var count: Int { get }
-
-  func contains(_ member: Element) -> Bool
-  func element(at i: Index) -> Element
-}
-
-/// An instance of this class has all `Set` data tail-allocated.
-/// Enough bytes are allocated to hold the bitmap for marking valid entries,
-/// keys, and values. The data layout starts with the bitmap, followed by the
-/// keys, followed by the values.
-//
-// See the docs at the top of the file for more details on this type
-//
-// NOTE: The precise layout of this type is relied on in the runtime
-// to provide a statically allocated empty singleton.
-// See stdlib/public/stubs/GlobalObjects.cpp for details.
-@_fixed_layout // FIXME(sil-serialize-all)
-@usableFromInline
-@_objc_non_lazy_realization
-internal class _RawSetStorage: _SwiftNativeNSSet {
-  /// The current number of occupied entries in this set.
-  @usableFromInline
-  @nonobjc
-  internal final var _count: Int
-
-  /// The maximum number of elements that can be inserted into this set without
-  /// exceeding the hash table's maximum load factor.
-  @usableFromInline
-  @nonobjc
-  internal final var _capacity: Int
-
-  /// The scale of this set. The number of buckets is 2 raised to the
-  /// power of `scale`.
-  @usableFromInline
-  @nonobjc
-  internal final var _scale: Int
-
-  @usableFromInline
-  internal final var _seed: Hasher._Seed
-
-  @usableFromInline
-  @nonobjc
-  internal final var _rawElements: UnsafeMutableRawPointer
-
-  // This type is made with allocWithTailElems, so no init is ever called.
-  // But we still need to have an init to satisfy the compiler.
-  @nonobjc
-  internal init(_doNotCallMe: ()) {
-    _sanityCheckFailure("This class cannot be directly initialized")
-  }
-
-  @inlinable
-  @nonobjc
-  internal final var _bucketCount: Int {
-    @inline(__always) get { return 1 &<< _scale }
-  }
-
-  @inlinable
-  @nonobjc
-  internal final var _metadata: UnsafeMutablePointer<_HashTable.Word> {
-    @inline(__always) get {
-      let address = Builtin.projectTailElems(self, _HashTable.Word.self)
-      return UnsafeMutablePointer(address)
-    }
-  }
-
-  // The _HashTable struct contains pointers into tail-allocated storage, so
-  // this is unsafe and needs `_fixLifetime` calls in the caller.
-  @inlinable
-  @nonobjc
-  internal final var _hashTable: _HashTable {
-    @inline(__always) get {
-      return _HashTable(words: _metadata, bucketCount: _bucketCount)
-    }
-  }
-}
-
-/// The storage class for the singleton empty set.
-/// The single instance of this class is created by the runtime.
-@_fixed_layout
-@usableFromInline
-internal class _EmptySetSingleton: _RawSetStorage {
-  @nonobjc
-  override internal init(_doNotCallMe: ()) {
-    _sanityCheckFailure("This class cannot be directly initialized")
-  }
-
-#if _runtime(_ObjC)
-  @objc
-  internal required init(objects: UnsafePointer<AnyObject?>, count: Int) {
-    _sanityCheckFailure("This class cannot be directly initialized")
-  }
-#endif
-}
-
-extension _RawSetStorage {
-  /// The empty singleton that is used for every single Set that is created
-  /// without any elements. The contents of the storage must never be mutated.
-  @inlinable
-  @nonobjc
-  internal static var empty: _EmptySetSingleton {
-    return Builtin.bridgeFromRawPointer(
-      Builtin.addressof(&_swiftEmptySetSingleton))
-  }
-}
-
-extension _EmptySetSingleton: _NSSetCore {
-#if _runtime(_ObjC)
-  //
-  // NSSet implementation, assuming Self is the empty singleton
-  //
-  @objc(copyWithZone:)
-  internal func copy(with zone: _SwiftNSZone?) -> AnyObject {
-    return self
-  }
-
-  @objc
-  internal var count: Int {
-    return 0
-  }
-
-  @objc(member:)
-  internal func member(_ object: AnyObject) -> AnyObject? {
-    return nil
-  }
-
-  @objc
-  internal func objectEnumerator() -> _NSEnumerator {
-    return _SwiftEmptyNSEnumerator()
-  }
-
-  @objc(countByEnumeratingWithState:objects:count:)
-  internal func countByEnumerating(
-    with state: UnsafeMutablePointer<_SwiftNSFastEnumerationState>,
-    objects: UnsafeMutablePointer<AnyObject>?, count: Int
-  ) -> Int {
-    // Even though we never do anything in here, we need to update the
-    // state so that callers know we actually ran.
-    var theState = state.pointee
-    if theState.state == 0 {
-      theState.state = 1 // Arbitrary non-zero value.
-      theState.itemsPtr = AutoreleasingUnsafeMutablePointer(objects)
-      theState.mutationsPtr = _fastEnumerationStorageMutationsPtr
-    }
-    state.pointee = theState
-    return 0
-  }
-#endif
-}
-
-// See the docs at the top of this file for a description of this type
-@_fixed_layout // FIXME(sil-serialize-all)
-@usableFromInline
-final internal class _SetStorage<Element: Hashable>
-  : _RawSetStorage, _NSSetCore {
-  // This type is made with allocWithTailElems, so no init is ever called.
-  // But we still need to have an init to satisfy the compiler.
-  @nonobjc
-  override internal init(_doNotCallMe: ()) {
-    _sanityCheckFailure("This class cannot be directly initialized")
-  }
-
-  deinit {
-    guard _count > 0 else { return }
-    if !_isPOD(Element.self) {
-      let elements = _elements
-      for index in _hashTable {
-        (elements + index.bucket).deinitialize(count: 1)
-      }
-    }
-    _fixLifetime(self)
-  }
-
-  @inlinable
-  final internal var _elements: UnsafeMutablePointer<Element> {
-    @inline(__always)
-    get {
-      return self._rawElements.assumingMemoryBound(to: Element.self)
-    }
-  }
-
-  internal var asNative: _NativeSet<Element> {
-    return _NativeSet(self)
-  }
-
-#if _runtime(_ObjC)
-  @objc
-  internal required init(objects: UnsafePointer<AnyObject?>, count: Int) {
-    _sanityCheckFailure("don't call this designated initializer")
-  }
-
-  @objc(copyWithZone:)
-  internal func copy(with zone: _SwiftNSZone?) -> AnyObject {
-    return self
-  }
-
-  @objc
-  internal var count: Int {
-    return _count
-  }
-
-  @objc
-  internal func objectEnumerator() -> _NSEnumerator {
-    return _SwiftSetNSEnumerator<Element>(asNative)
-  }
-
-  @objc(countByEnumeratingWithState:objects:count:)
-  internal func countByEnumerating(
-    with state: UnsafeMutablePointer<_SwiftNSFastEnumerationState>,
-    objects: UnsafeMutablePointer<AnyObject>?, count: Int
-  ) -> Int {
-    var theState = state.pointee
-    if theState.state == 0 {
-      theState.state = 1 // Arbitrary non-zero value.
-      theState.itemsPtr = AutoreleasingUnsafeMutablePointer(objects)
-      theState.mutationsPtr = _fastEnumerationStorageMutationsPtr
-      theState.extra.0 = CUnsignedLong(asNative.startIndex.bucket)
-    }
-
-    // Test 'objects' rather than 'count' because (a) this is very rare anyway,
-    // and (b) the optimizer should then be able to optimize away the
-    // unwrapping check below.
-    if _slowPath(objects == nil) {
-      return 0
-    }
-
-    let unmanagedObjects = _UnmanagedAnyObjectArray(objects!)
-    var index = _HashTable.Index(bucket: Int(theState.extra.0))
-    let endIndex = asNative.endIndex
-    _precondition(index == endIndex || _hashTable.isValid(index))
-    var stored = 0
-    for i in 0..<count {
-      if index == endIndex { break }
-      let element = _elements[index.bucket]
-      unmanagedObjects[i] = _bridgeAnythingToObjectiveC(element)
-      stored += 1
-      index = asNative.index(after: index)
-    }
-    theState.extra.0 = CUnsignedLong(index.bucket)
-    state.pointee = theState
-    return stored
-  }
-
-  @objc(member:)
-  internal func member(_ object: AnyObject) -> AnyObject? {
-    guard let native = _conditionallyBridgeFromObjectiveC(object, Element.self)
-    else { return nil }
-
-    let (index, found) = asNative.find(native)
-    guard found else { return nil }
-    return _bridgeAnythingToObjectiveC(_elements[index.bucket])
-  }
-#endif
-}
-
-extension _SetStorage {
-  @usableFromInline
-  @_effects(releasenone)
-  internal static func reallocate(
-    original: _RawSetStorage,
-    capacity: Int
-  ) -> (storage: _SetStorage, rehash: Bool) {
-    _sanityCheck(capacity >= original._count)
-    let scale = _HashTable.scale(forCapacity: capacity)
-    let rehash = (scale != original._scale)
-    let newStorage = _SetStorage<Element>.allocate(scale: scale)
-    return (newStorage, rehash)
-  }
-
-  @usableFromInline
-  @_effects(releasenone)
-  static internal func allocate(capacity: Int) -> _SetStorage {
-    let scale = _HashTable.scale(forCapacity: capacity)
-    return allocate(scale: scale)
-  }
-
-  static internal func allocate(scale: Int) -> _SetStorage {
-    // The entry count must be representable by an Int value; hence the scale's
-    // peculiar upper bound.
-    _sanityCheck(scale >= 0 && scale < Int.bitWidth - 1)
-
-    let bucketCount = 1 &<< scale
-    let wordCount = _UnsafeBitset.wordCount(forCapacity: bucketCount)
-    let storage = Builtin.allocWithTailElems_2(
-      _SetStorage<Element>.self,
-      wordCount._builtinWordValue, _HashTable.Word.self,
-      bucketCount._builtinWordValue, Element.self)
-
-    let metadataAddr = Builtin.projectTailElems(storage, _HashTable.Word.self)
-    let elementsAddr = Builtin.getTailAddr_Word(
-      metadataAddr, wordCount._builtinWordValue, _HashTable.Word.self,
-      Element.self)
-    storage._count = 0
-    storage._capacity = _HashTable.capacity(forScale: scale)
-    storage._scale = scale
-    storage._rawElements = UnsafeMutableRawPointer(elementsAddr)
-
-    // We use a slightly different hash seed whenever we change the size of the
-    // hash table, so that we avoid certain copy operations becoming quadratic,
-    // without breaking value semantics. (For background details, see
-    // https://bugs.swift.org/browse/SR-3268)
-
-    // FIXME: Use true per-instance seeding instead. Per-capacity seeding still
-    // leaves hash values the same in same-sized tables, which may affect
-    // operations on two tables at once. (E.g., union.)
-    storage._seed = (
-      Hasher._seed.0 ^ UInt64(truncatingIfNeeded: scale),
-      Hasher._seed.1)
-
-    // Initialize hash table metadata.
-    storage._hashTable.clear()
-    return storage
-  }
-}
-
-/// A wrapper around _RawSetStorage that provides most of the
-/// implementation of Set.
-@usableFromInline
-@_fixed_layout
-internal struct _NativeSet<Element: Hashable> {
-  /// See the comments on _RawSetStorage and its subclasses to understand why we
-  /// store an untyped storage here.
-  @usableFromInline
-  internal var _storage: _RawSetStorage
-
-  /// Constructs an instance from the empty singleton.
-  @inlinable
-  @inline(__always)
-  internal init() {
-    self._storage = _RawSetStorage.empty
-  }
-
-  /// Constructs a native set adopting the given storage.
-  @inlinable
-  @inline(__always)
-  internal init(_ storage: _RawSetStorage) {
-    self._storage = storage
-  }
-
-  @usableFromInline
-  @_effects(releasenone)
-  internal init(capacity: Int) {
-    let scale = _HashTable.scale(forCapacity: capacity)
-    self._storage = _SetStorage<Element>.allocate(scale: scale)
-  }
-
-#if _runtime(_ObjC)
-  @inlinable
-  internal init(_ cocoa: _CocoaSet) {
-    self.init(cocoa, capacity: cocoa.count)
-  }
-
-  @inlinable
-  internal init(_ cocoa: _CocoaSet, capacity: Int) {
-    _sanityCheck(cocoa.count <= capacity)
-    self.init(capacity: capacity)
-    for element in cocoa {
-      let nativeElement = _forceBridgeFromObjectiveC(element, Element.self)
-      insertNew(nativeElement, isUnique: true)
-    }
-  }
-#endif
-}
-
-extension _NativeSet { // Primitive fields
-  @inlinable
-  internal var capacity: Int {
-    @inline(__always)
-    get {
-      return _assumeNonNegative(_storage._capacity)
-    }
-  }
-
-  @inlinable
-  internal var hashTable: _HashTable {
-    @inline(__always) get {
-      return _storage._hashTable
-    }
-  }
-
-  // This API is unsafe and needs a `_fixLifetime` in the caller.
-  @inlinable
-  internal var _elements: UnsafeMutablePointer<Element> {
-    return _storage._rawElements.assumingMemoryBound(to: Element.self)
-  }
-}
-
-extension _NativeSet { // Low-level unchecked operations
-  @inlinable
-  @inline(__always)
-  internal func uncheckedElement(at index: Index) -> Element {
-    defer { _fixLifetime(self) }
-    _sanityCheck(hashTable.isOccupied(index))
-    return _elements[index.bucket]
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func uncheckedInitialize(at index: Index, to element: Element) {
-    _sanityCheck(hashTable.isValid(index))
-    (_elements + index.bucket).initialize(to: element)
-  }
-}
-
-extension _NativeSet { // Low-level lookup operations
-  @inlinable
-  @inline(__always)
-  internal func hashValue(for element: Element) -> Int {
-    return element._rawHashValue(seed: _storage._seed)
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func find(_ element: Element) -> (index: Index, found: Bool) {
-    return find(element, hashValue: self.hashValue(for: element))
-  }
-
-  /// Search for a given element, assuming it has the specified hash value.
-  ///
-  /// If the element is not present in this set, return the position where it
-  /// could be inserted.
-  @inlinable
-  @inline(__always)
-  internal func find(
-    _ element: Element,
-    hashValue: Int
-  ) -> (index: Index, found: Bool) {
-    let hashTable = self.hashTable
-    var index = hashTable.idealIndex(forHashValue: hashValue)
-    while hashTable._isOccupied(index) {
-      if uncheckedElement(at: index) == element {
-        return (index, true)
-      }
-      index = hashTable.index(wrappedAfter: index)
-    }
-    return (index, false)
-  }
-}
-
-extension _NativeSet { // ensureUnique
-  @inlinable
-  internal mutating func resize(capacity: Int) {
-    let capacity = Swift.max(capacity, self.capacity)
-    let result = _NativeSet(_SetStorage<Element>.allocate(capacity: capacity))
-    if count > 0 {
-      for index in hashTable {
-        let element = (self._elements + index.bucket).move()
-        result._unsafeInsertNew(element)
-      }
-      // Clear out old storage, ensuring that its deinit won't overrelease the
-      // elements we've just moved out.
-      _storage._hashTable.clear()
-      _storage._count = 0
-    }
-    _storage = result._storage
-  }
-
-  @inlinable
-  internal mutating func copy(capacity: Int) -> Bool {
-    let capacity = Swift.max(capacity, self.capacity)
-    let (newStorage, rehash) = _SetStorage<Element>.reallocate(
-      original: _storage,
-      capacity: capacity)
-    let result = _NativeSet(newStorage)
-    if count > 0 {
-      if rehash {
-        for index in hashTable {
-          result._unsafeInsertNew(self.uncheckedElement(at: index))
-        }
-      } else {
-        result.hashTable.copyContents(of: hashTable)
-        result._storage._count = self.count
-        for index in hashTable {
-          let element = uncheckedElement(at: index)
-          result.uncheckedInitialize(at: index, to: element)
-        }
-      }
-    }
-    _storage = result._storage
-    return rehash
-  }
-
-  /// Ensure storage of self is uniquely held and can hold at least `capacity`
-  /// elements. Returns true iff contents were rehashed.
-  @inlinable
-  @inline(__always)
-  internal mutating func ensureUnique(isUnique: Bool, capacity: Int) -> Bool {
-    if _fastPath(capacity <= self.capacity && isUnique) {
-      return false
-    }
-    guard isUnique else {
-      return copy(capacity: capacity)
-    }
-    resize(capacity: capacity)
-    return true
-  }
-
-  @inlinable
-  internal mutating func reserveCapacity(_ capacity: Int, isUnique: Bool) {
-    _ = ensureUnique(isUnique: isUnique, capacity: capacity)
-  }
-}
-
-extension _NativeSet: _SetBuffer {
-  @usableFromInline
-  internal typealias Index = _HashTable.Index
-
-  @inlinable
-  internal var startIndex: Index {
-    return hashTable.startIndex
-  }
-
-  @inlinable
-  internal var endIndex: Index {
-    return hashTable.endIndex
-  }
-
-  @inlinable
-  internal func index(after index: Index) -> Index {
-    return hashTable.index(after: index)
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func index(for element: Element) -> Index? {
-    if count == 0 {
-      // Fast path that avoids computing the hash of the key.
-      return nil
-    }
-    let (index, found) = find(element)
-    return found ? index : nil
-  }
-
-  @inlinable
-  internal var count: Int {
-    @inline(__always) get {
-      return _assumeNonNegative(_storage._count)
-    }
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func contains(_ member: Element) -> Bool {
-    // Fast path: Don't calculate the hash if the set has no elements.
-    if count == 0 { return false }
-    return find(member).found
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func element(at index: Index) -> Element {
-    hashTable.checkOccupied(index)
-    return _elements[index.bucket]
-  }
-}
-
-#if _runtime(_ObjC)
-extension _NativeSet { // Bridging
-  @usableFromInline
-  internal func bridged() -> _NSSet {
-    // We can zero-cost bridge if our keys are verbatim
-    // or if we're the empty singleton.
-
-    // Temporary var for SOME type safety before a cast.
-    let nsSet: _NSSetCore
-
-    if _storage === _RawSetStorage.empty || count == 0 {
-      nsSet = _RawSetStorage.empty
-    } else if _isBridgedVerbatimToObjectiveC(Element.self) {
-      nsSet = unsafeDowncast(_storage, to: _SetStorage<Element>.self)
-    } else {
-      nsSet = _SwiftDeferredNSSet(self)
-    }
-
-    // Cast from "minimal NSSet" to "NSSet"
-    // Note that if you actually ask Swift for this cast, it will fail.
-    // Never trust a shadow protocol!
-    return unsafeBitCast(nsSet, to: _NSSet.self)
-  }
-}
-#endif
-
-// This function has a highly visible name to make it stand out in stack traces.
-@usableFromInline
-@inline(never)
-internal func ELEMENT_TYPE_OF_SET_VIOLATES_HASHABLE_REQUIREMENTS(
-  _ elementType: Any.Type
-) -> Never {
-  _assertionFailure(
-    "Fatal error",
-    """
-    Duplicate elements of type '\(elementType)' were found in a Set.
-    This usually means either that the type violates Hashable's requirements, or
-    that members of such a set were mutated after insertion.
-    """,
-    flags: _fatalErrorFlags())
-}
-
-extension _NativeSet { // Insertions
-  /// Insert a new element into uniquely held storage.
-  /// Storage must be uniquely referenced with adequate capacity.
-  /// The `element` must not be already present in the Set.
-  @inlinable
-  internal func _unsafeInsertNew(_ element: Element) {
-    _sanityCheck(count + 1 <= capacity)
-    let hashValue = self.hashValue(for: element)
-    if _isDebugAssertConfiguration() {
-      // In debug builds, perform a full lookup and trap if we detect duplicate
-      // elements -- these imply that the Element type violates Hashable
-      // requirements. This is generally more costly than a direct insertion,
-      // because we'll need to compare elements in case of hash collisions.
-      let (index, found) = find(element, hashValue: hashValue)
-      guard !found else {
-        ELEMENT_TYPE_OF_SET_VIOLATES_HASHABLE_REQUIREMENTS(Element.self)
-      }
-      hashTable.insert(index)
-      uncheckedInitialize(at: index, to: element)
-    } else {
-      let index = hashTable.insertNew(hashValue: hashValue)
-      uncheckedInitialize(at: index, to: element)
-    }
-    _storage._count += 1
-  }
-
-  /// Insert a new element into uniquely held storage.
-  /// Storage must be uniquely referenced.
-  /// The `element` must not be already present in the Set.
-  @inlinable
-  internal mutating func insertNew(_ element: Element, isUnique: Bool) {
-    _ = ensureUnique(isUnique: isUnique, capacity: count + 1)
-    _unsafeInsertNew(element)
-  }
-
-  @inlinable
-  internal func _unsafeInsertNew(_ element: Element, at index: Index) {
-    hashTable.insert(index)
-    uncheckedInitialize(at: index, to: element)
-    _storage._count += 1
-  }
-
-  @inlinable
-  internal mutating func insertNew(
-    _ element: Element,
-    at index: Index,
-    isUnique: Bool
-  ) {
-    _sanityCheck(!hashTable.isOccupied(index))
-    var index = index
-    if ensureUnique(isUnique: isUnique, capacity: count + 1) {
-      let (i, f) = find(element)
-      if f {
-        ELEMENT_TYPE_OF_SET_VIOLATES_HASHABLE_REQUIREMENTS(Element.self)
-      }
-      index = i
-    }
-    _unsafeInsertNew(element, at: index)
-  }
-
-  @inlinable
-  internal mutating func update(
-    with element: Element,
-    isUnique: Bool
-  ) -> Element? {
-    var (index, found) = find(element)
-    let rehashed = ensureUnique(
-      isUnique: isUnique,
-      capacity: count + (found ? 0 : 1))
-    if rehashed {
-      let (i, f) = find(element)
-      if f != found {
-        ELEMENT_TYPE_OF_SET_VIOLATES_HASHABLE_REQUIREMENTS(Element.self)
-      }
-      index = i
-    }
-    if found {
-      let old = (_elements + index.bucket).move()
-      uncheckedInitialize(at: index, to: element)
-      return old
-    }
-    _unsafeInsertNew(element, at: index)
-    return nil
-  }
-}
-
-extension _NativeSet: _HashTableDelegate {
-  @inlinable
-  @inline(__always)
-  internal func hashValue(at index: Index) -> Int {
-    return hashValue(for: uncheckedElement(at: index))
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func moveEntry(from source: Index, to target: Index) {
-    (_elements + target.bucket)
-      .moveInitialize(from: _elements + source.bucket, count: 1)
-  }
-}
-
-extension _NativeSet { // Deletion
-  @inlinable
-  internal mutating func _delete(at index: Index) {
-    hashTable.delete(at: index, with: self)
-    _storage._count -= 1
-  }
-
-  @inlinable
-  @inline(__always)
-  internal mutating func uncheckedRemove(
-    at index: Index,
-    isUnique: Bool) -> Element {
-    _sanityCheck(hashTable.isOccupied(index))
-    let rehashed = ensureUnique(isUnique: isUnique, capacity: capacity)
-    _sanityCheck(!rehashed)
-    let old = (_elements + index.bucket).move()
-    _delete(at: index)
-    return old
-  }
-
-  @inlinable
-  @inline(__always)
-  internal mutating func remove(at index: Index, isUnique: Bool) -> Element {
-    _precondition(hashTable.isOccupied(index), "Invalid index")
-    return uncheckedRemove(at: index, isUnique: isUnique)
-  }
-
-  @usableFromInline
-  internal mutating func removeAll(isUnique: Bool) {
-    guard isUnique else {
-      let scale = self._storage._scale
-      _storage = _SetStorage<Element>.allocate(scale: scale)
-      return
-    }
-    for index in hashTable {
-      (_elements + index.bucket).deinitialize(count: 1)
-    }
-    hashTable.clear()
-    _storage._count = 0
-  }
-}
-
-#if _runtime(_ObjC)
-/// An NSEnumerator that works with any _NativeSet of verbatim bridgeable
-/// elements. Used by the various NSSet impls.
-final internal class _SwiftSetNSEnumerator<Element: Hashable>
-  : _SwiftNativeNSEnumerator, _NSEnumerator {
-
-  @nonobjc internal var base: _NativeSet<Element>
-  @nonobjc internal var bridgedElements: _BridgingHashBuffer?
-  @nonobjc internal var nextIndex: _NativeSet<Element>.Index
-  @nonobjc internal var endIndex: _NativeSet<Element>.Index
-
-  @objc
-  internal override required init() {
-    _sanityCheckFailure("don't call this designated initializer")
-  }
-
-  internal init(_ base: _NativeSet<Element>) {
-    _sanityCheck(_isBridgedVerbatimToObjectiveC(Element.self))
-    self.base = base
-    self.bridgedElements = nil
-    self.nextIndex = base.startIndex
-    self.endIndex = base.endIndex
-  }
-
-  @nonobjc
-  internal init(_ deferred: _SwiftDeferredNSSet<Element>) {
-    _sanityCheck(!_isBridgedVerbatimToObjectiveC(Element.self))
-    self.base = deferred.native
-    self.bridgedElements = deferred.bridgeElements()
-    self.nextIndex = base.startIndex
-    self.endIndex = base.endIndex
-  }
-
-  private func bridgedElement(at index: _HashTable.Index) -> AnyObject {
-    _sanityCheck(base.hashTable.isOccupied(index))
-    if let bridgedElements = self.bridgedElements {
-      return bridgedElements[index]
-    }
-    return _bridgeAnythingToObjectiveC(base.element(at: index))
-  }
-
-  //
-  // NSEnumerator implementation.
-  //
-  // Do not call any of these methods from the standard library!
-  //
-
-  @objc
-  internal func nextObject() -> AnyObject? {
-    if nextIndex == endIndex {
-      return nil
-    }
-    let index = nextIndex
-    nextIndex = base.index(after: nextIndex)
-    return self.bridgedElement(at: index)
-  }
-
-  @objc(countByEnumeratingWithState:objects:count:)
-  internal func countByEnumerating(
-    with state: UnsafeMutablePointer<_SwiftNSFastEnumerationState>,
-    objects: UnsafeMutablePointer<AnyObject>,
-    count: Int
-  ) -> Int {
-    var theState = state.pointee
-    if theState.state == 0 {
-      theState.state = 1 // Arbitrary non-zero value.
-      theState.itemsPtr = AutoreleasingUnsafeMutablePointer(objects)
-      theState.mutationsPtr = _fastEnumerationStorageMutationsPtr
-    }
-
-    if nextIndex == endIndex {
-      state.pointee = theState
-      return 0
-    }
-
-    // Return only a single element so that code can start iterating via fast
-    // enumeration, terminate it, and continue via NSEnumerator.
-    let unmanagedObjects = _UnmanagedAnyObjectArray(objects)
-    unmanagedObjects[0] = self.bridgedElement(at: nextIndex)
-    nextIndex = base.index(after: nextIndex)
-    state.pointee = theState
-    return 1
-  }
-}
-#endif
-
-#if _runtime(_ObjC)
-/// This class exists for Objective-C bridging. It holds a reference to a
-/// _NativeSet, and can be upcast to NSSelf when bridging is necessary.  This is
-/// the fallback implementation for situations where toll-free bridging isn't
-/// possible. On first access, a _NativeSet of AnyObject will be constructed
-/// containing all the bridged elements.
-final internal class _SwiftDeferredNSSet<Element: Hashable>
-  : _SwiftNativeNSSet, _NSSetCore {
-
-  // This stored property must be stored at offset zero.  We perform atomic
-  // operations on it.
-  //
-  // Do not access this property directly.
-  @nonobjc
-  private var _bridgedElements_DoNotUse: AnyObject?
-
-  /// The unbridged elements.
-  internal var native: _NativeSet<Element>
-
-  internal init(_ native: _NativeSet<Element>) {
-    _sanityCheck(native.count > 0)
-    _sanityCheck(!_isBridgedVerbatimToObjectiveC(Element.self))
-    self.native = native
-    super.init()
-  }
-
-  /// Returns the pointer to the stored property, which contains bridged
-  /// Set elements.
-  @nonobjc
-  private var _bridgedElementsPtr: UnsafeMutablePointer<AnyObject?> {
-    return _getUnsafePointerToStoredProperties(self)
-      .assumingMemoryBound(to: Optional<AnyObject>.self)
-  }
-
-  /// The buffer for bridged Set elements, if present.
-  @nonobjc
-  private var _bridgedElements: _BridgingHashBuffer? {
-    guard let ref = _stdlib_atomicLoadARCRef(object: _bridgedElementsPtr) else {
-      return nil
-    }
-    return unsafeDowncast(ref, to: _BridgingHashBuffer.self)
-  }
-
-  /// Attach a buffer for bridged Set elements.
-  @nonobjc
-  private func _initializeBridgedElements(_ storage: _BridgingHashBuffer) {
-    _stdlib_atomicInitializeARCRef(
-      object: _bridgedElementsPtr,
-      desired: storage)
-  }
-
-  @nonobjc
-  internal func bridgeElements() -> _BridgingHashBuffer {
-    if let bridgedElements = _bridgedElements { return bridgedElements }
-
-    // Allocate and initialize heap storage for bridged objects.
-    let bridged = _BridgingHashBuffer.allocate(
-      owner: native._storage,
-      hashTable: native.hashTable)
-    for index in native.hashTable {
-      let object = _bridgeAnythingToObjectiveC(native.element(at: index))
-      bridged.initialize(at: index, to: object)
-    }
-
-    // Atomically put the bridged elements in place.
-    _initializeBridgedElements(bridged)
-    return _bridgedElements!
-  }
-
-  @objc
-  internal required init(objects: UnsafePointer<AnyObject?>, count: Int) {
-    _sanityCheckFailure("don't call this designated initializer")
-  }
-
-  @objc(copyWithZone:)
-  internal func copy(with zone: _SwiftNSZone?) -> AnyObject {
-    // Instances of this class should be visible outside of standard library as
-    // having `NSSet` type, which is immutable.
-    return self
-  }
-
-  @objc(member:)
-  internal func member(_ object: AnyObject) -> AnyObject? {
-    guard let element = _conditionallyBridgeFromObjectiveC(object, Element.self)
-    else { return nil }
-
-    let (index, found) = native.find(element)
-    guard found else { return nil }
-    let bridged = bridgeElements()
-    return bridged[index]
-  }
-
-  @objc
-  internal func objectEnumerator() -> _NSEnumerator {
-    return _SwiftSetNSEnumerator<Element>(self)
-  }
-
-  @objc
-  internal var count: Int {
-    return native.count
-  }
-
-  @objc(countByEnumeratingWithState:objects:count:)
-  internal func countByEnumerating(
-    with state: UnsafeMutablePointer<_SwiftNSFastEnumerationState>,
-    objects: UnsafeMutablePointer<AnyObject>?,
-    count: Int
-  ) -> Int {
-    var theState = state.pointee
-    if theState.state == 0 {
-      theState.state = 1 // Arbitrary non-zero value.
-      theState.itemsPtr = AutoreleasingUnsafeMutablePointer(objects)
-      theState.mutationsPtr = _fastEnumerationStorageMutationsPtr
-      theState.extra.0 = CUnsignedLong(native.startIndex.bucket)
-    }
-
-    // Test 'objects' rather than 'count' because (a) this is very rare anyway,
-    // and (b) the optimizer should then be able to optimize away the
-    // unwrapping check below.
-    if _slowPath(objects == nil) {
-      return 0
-    }
-
-    let unmanagedObjects = _UnmanagedAnyObjectArray(objects!)
-    var index = _NativeSet<Element>.Index(bucket: Int(theState.extra.0))
-    let endIndex = native.endIndex
-    _precondition(index == endIndex || native.hashTable.isValid(index))
-
-    // Only need to bridge once, so we can hoist it out of the loop.
-    let bridgedElements = bridgeElements()
-
-    var stored = 0
-    for i in 0..<count {
-      if index == endIndex { break }
-      unmanagedObjects[i] = bridgedElements[index]
-      stored += 1
-      index = native.index(after: index)
-    }
-    theState.extra.0 = CUnsignedLong(index.bucket)
-    state.pointee = theState
-    return stored
-  }
-}
-#else
-// FIXME: Remove this.
-final internal class _SwiftDeferredNSSet<Element: Hashable> { }
-#endif
-
-#if _runtime(_ObjC)
-@usableFromInline
-@_fixed_layout
-internal struct _CocoaSet {
-  @usableFromInline
-  internal let object: _NSSet
-
-  @inlinable
-  internal init(_ object: _NSSet) {
-    self.object = object
-  }
-}
-
-extension _CocoaSet {
-  @usableFromInline
-  @_effects(releasenone)
-  internal func member(for index: Index) -> AnyObject {
-    return index.allKeys[index.currentKeyIndex]
-  }
-
-  @inlinable
-  internal func member(for element: AnyObject) -> AnyObject? {
-    return object.member(element)
-  }
-}
-
-extension _CocoaSet: Equatable {
-  @usableFromInline
-  internal static func ==(lhs: _CocoaSet, rhs: _CocoaSet) -> Bool {
-    return _stdlib_NSObject_isEqual(lhs.object, rhs.object)
-  }
-}
-
-extension _CocoaSet: _SetBuffer {
-  @usableFromInline
-  internal typealias Element = AnyObject
-
-  @inlinable
-  internal var startIndex: Index {
-    return Index(self, startIndex: ())
-  }
-
-  @inlinable
-  internal var endIndex: Index {
-    return Index(self, endIndex: ())
-  }
-
-  @inlinable
-  internal func index(after i: Index) -> Index {
-    var i = i
-    formIndex(after: &i)
-    return i
-  }
-
-  @usableFromInline
-  @_effects(releasenone)
-  internal func formIndex(after i: inout Index) {
-    _precondition(i.base.object === self.object, "Invalid index")
-    _precondition(i.currentKeyIndex < i.allKeys.value,
-      "Cannot increment endIndex")
-    i.currentKeyIndex += 1
-  }
-
-  @usableFromInline
-  internal func index(for element: AnyObject) -> Index? {
-    // Fast path that does not involve creating an array of all keys.  In case
-    // the key is present, this lookup is a penalty for the slow path, but the
-    // potential savings are significant: we could skip a memory allocation and
-    // a linear search.
-    if !contains(element) {
-      return nil
-    }
-
-    let allKeys = _stdlib_NSSet_allObjects(object)
-    var keyIndex = -1
-    for i in 0..<allKeys.value {
-      if _stdlib_NSObject_isEqual(element, allKeys[i]) {
-        keyIndex = i
-        break
-      }
-    }
-    _sanityCheck(keyIndex >= 0,
-        "Key was found in fast path, but not found later?")
-    return Index(self, allKeys, keyIndex)
-  }
-
-  @inlinable
-  internal var count: Int {
-    return object.count
-  }
-
-  @inlinable
-  internal func contains(_ element: AnyObject) -> Bool {
-    return object.member(element) != nil
-  }
-
-  @usableFromInline
-  internal func element(at i: Index) -> AnyObject {
-    let value: AnyObject? = i.allKeys[i.currentKeyIndex]
-    _sanityCheck(value != nil, "Item not found in underlying NSSet")
-    return value!
-  }
-}
-#endif
-
-extension Set {
-  @usableFromInline
-  @_frozen
-  internal enum _Variant {
-    case native(_NativeSet<Element>)
-#if _runtime(_ObjC)
-    case cocoa(_CocoaSet)
-#endif
-  }
-}
-
-extension Set._Variant {
-#if _runtime(_ObjC)
-  @usableFromInline
-  @_transparent
-  internal var guaranteedNative: Bool {
-    return _canBeClass(Element.self) == 0
-  }
-
-  /// Allow the optimizer to consider the surrounding code unreachable if
-  /// Set<Element> is guaranteed to be native.
-  @usableFromInline
-  @_transparent
-  internal func cocoaPath() {
-    if guaranteedNative {
-      _conditionallyUnreachable()
-    }
-  }
-#endif
-
-  @inlinable
-  internal mutating func isUniquelyReferenced() -> Bool {
-    // Note that &self drills down through .native(_NativeSet) to the first
-    // property in _NativeSet, which is the reference to the storage.
-    switch self {
-    case .native:
-      return _isUnique_native(&self)
-#if _runtime(_ObjC)
-    case .cocoa:
-      cocoaPath()
-      // Don't consider Cocoa buffer mutable, even if it is mutable and is
-      // uniquely referenced.
-      return false
-#endif
-    }
-  }
-
-  @usableFromInline @_transparent
-  internal var asNative: _NativeSet<Element> {
-    get {
-      switch self {
-      case .native(let nativeSet):
-        return nativeSet
-#if _runtime(_ObjC)
-      case .cocoa:
-        _sanityCheckFailure("internal error: not backed by native buffer")
-#endif
-      }
-    }
-    set {
-      self = .native(newValue)
-    }
-  }
-
-#if _runtime(_ObjC)
-  @inlinable
-  internal var asCocoa: _CocoaSet {
-    switch self {
-    case .native:
-      _sanityCheckFailure("internal error: not backed by NSSet")
-    case .cocoa(let cocoa):
-      return cocoa
-    }
-  }
-#endif
-
-  /// Reserves enough space for the specified number of elements to be stored
-  /// without reallocating additional storage.
-  @inlinable
-  internal mutating func reserveCapacity(_ capacity: Int) {
-    switch self {
-    case .native:
-      let isUnique = isUniquelyReferenced()
-      asNative.reserveCapacity(capacity, isUnique: isUnique)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      let capacity = Swift.max(cocoa.count, capacity)
-      self = .native(_NativeSet(cocoa, capacity: capacity))
-#endif
-    }
-  }
-
-  /// The number of elements that can be stored without expanding the current
-  /// storage.
-  ///
-  /// For bridged storage, this is equal to the current count of the
-  /// collection, since any addition will trigger a copy of the elements into
-  /// newly allocated storage. For native storage, this is the element count
-  /// at which adding any more elements will exceed the load factor.
-  @inlinable
-  internal var capacity: Int {
-    switch self {
-    case .native:
-      return asNative.capacity
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      return cocoa.count
-#endif
-    }
-  }
-}
-
-extension Set._Variant: _SetBuffer {
-  @usableFromInline
-  internal typealias Index = Set<Element>.Index
-
-  @inlinable
-  internal var startIndex: Index {
-    switch self {
-    case .native:
-      return Index(_native: asNative.startIndex)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoaSet):
-      cocoaPath()
-      return Index(_cocoa: cocoaSet.startIndex)
-#endif
-    }
-  }
-
-  @inlinable
-  internal var endIndex: Index {
-    switch self {
-    case .native:
-      return Index(_native: asNative.endIndex)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoaSet):
-      cocoaPath()
-      return Index(_cocoa: cocoaSet.endIndex)
-#endif
-    }
-  }
-
-  @inlinable
-  internal func index(after i: Index) -> Index {
-    switch self {
-    case .native:
-      return Index(_native: asNative.index(after: i._asNative))
-#if _runtime(_ObjC)
-    case .cocoa(let cocoaSet):
-      cocoaPath()
-      return Index(_cocoa: cocoaSet.index(after: i._asCocoa))
-#endif
-    }
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func index(for element: Element) -> Index? {
-    switch self {
-    case .native:
-      guard let index = asNative.index(for: element) else { return nil }
-      return Index(_native: index)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      let cocoaElement = _bridgeAnythingToObjectiveC(element)
-      guard let index = cocoa.index(for: cocoaElement) else { return nil }
-      return Index(_cocoa: index)
-#endif
-    }
-  }
-
-  @inlinable
-  internal var count: Int {
-    @inline(__always)
-    get {
-      switch self {
-      case .native:
-        return asNative.count
-#if _runtime(_ObjC)
-      case .cocoa(let cocoa):
-        cocoaPath()
-        return cocoa.count
-#endif
-      }
-    }
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func contains(_ member: Element) -> Bool {
-    switch self {
-    case .native:
-      return asNative.contains(member)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      return cocoa.contains(_bridgeAnythingToObjectiveC(member))
-#endif
-    }
-  }
-
-  @inlinable
-  @inline(__always)
-  internal func element(at i: Index) -> Element {
-    switch self {
-    case .native:
-      return asNative.element(at: i._asNative)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      let cocoaMember = cocoa.element(at: i._asCocoa)
-      return _forceBridgeFromObjectiveC(cocoaMember, Element.self)
-#endif
-    }
-  }
-}
-
-extension Set._Variant {
-  @inlinable
-  internal mutating func update(with value: Element) -> Element? {
-    switch self {
-    case .native:
-      let isUnique = self.isUniquelyReferenced()
-      return asNative.update(with: value, isUnique: isUnique)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      // Make sure we have space for an extra element.
-      var native = _NativeSet<Element>(cocoa, capacity: cocoa.count + 1)
-      let old = native.update(with: value, isUnique: true)
-      self = .native(native)
-      return old
-#endif
-    }
-  }
-
-  @inlinable
-  internal mutating func insert(
-    _ element: Element
-  ) -> (inserted: Bool, memberAfterInsert: Element) {
-    switch self {
-    case .native:
-      let (index, found) = asNative.find(element)
-      if found {
-        return (false, asNative.uncheckedElement(at: index))
-      }
-      let isUnique = self.isUniquelyReferenced()
-      asNative.insertNew(element, at: index, isUnique: isUnique)
-      return (true, element)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      // Make sure we have space for an extra element.
-      let cocoaMember = _bridgeAnythingToObjectiveC(element)
-      if let m = cocoa.member(for: cocoaMember) {
-        return (false, _forceBridgeFromObjectiveC(m, Element.self))
-      }
-      var native = _NativeSet<Element>(cocoa, capacity: cocoa.count + 1)
-      native.insertNew(element, isUnique: true)
-      self = .native(native)
-      return (true, element)
-#endif
-    }
-  }
-
-  @inlinable
-  @discardableResult
-  internal mutating func remove(at index: Index) -> Element {
-    switch self {
-    case .native:
-      let isUnique = isUniquelyReferenced()
-      return asNative.remove(at: index._asNative, isUnique: isUnique)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      // We have to migrate the data first.  But after we do so, the Cocoa
-      // index becomes useless, so get the element first.
-      let cocoaMember = cocoa.member(for: index._asCocoa)
-      let nativeMember = _forceBridgeFromObjectiveC(cocoaMember, Element.self)
-      return _migrateToNative(cocoa, removing: nativeMember)
-#endif
-    }
-  }
-
-  @inlinable
-  @discardableResult
-  internal mutating func remove(_ member: Element) -> Element? {
-    switch self {
-    case .native:
-      let (index, found) = asNative.find(member)
-      guard found else { return nil }
-      let isUnique = isUniquelyReferenced()
-      return asNative.uncheckedRemove(at: index, isUnique: isUnique)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      let cocoaMember = _bridgeAnythingToObjectiveC(member)
-      guard cocoa.contains(cocoaMember) else { return nil }
-      return _migrateToNative(cocoa, removing: member)
-#endif
-    }
-  }
-
-#if _runtime(_ObjC)
-  @inlinable
-  internal mutating func _migrateToNative(
-    _ cocoa: _CocoaSet,
-    removing member: Element
-  ) -> Element {
-    // FIXME(performance): fuse data migration and element deletion into one
-    // operation.
-    var native = _NativeSet<Element>(cocoa)
-    let (index, found) = native.find(member)
-    _precondition(found, "Bridging did not preserve equality")
-    let old = native.remove(at: index, isUnique: true)
-    _precondition(member == old, "Bridging did not preserve equality")
-    self = .native(native)
-    return old
-  }
-#endif
-
-  @inlinable
-  internal mutating func removeAll(keepingCapacity keepCapacity: Bool) {
-    if !keepCapacity {
-      self = .native(_NativeSet<Element>())
-      return
-    }
-    guard count > 0 else { return }
-
-    switch self {
-    case .native:
-      let isUnique = isUniquelyReferenced()
-      asNative.removeAll(isUnique: isUnique)
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      self = .native(_NativeSet(capacity: cocoa.count))
-#endif
-    }
-  }
-}
-
-extension Set._Variant {
-  /// Returns an iterator over the elements.
-  ///
-  /// - Complexity: O(1).
-  @inlinable
-  @inline(__always)
-  __consuming internal func makeIterator() -> Set<Element>.Iterator {
-    switch self {
-    case .native(let native):
-      return Set.Iterator(_native: native.makeIterator())
-#if _runtime(_ObjC)
-    case .cocoa(let cocoa):
-      cocoaPath()
-      return Set.Iterator(_cocoa: cocoa.makeIterator())
-#endif
-    }
-  }
-}
-
-#if _runtime(_ObjC)
-extension _CocoaSet {
-  @_fixed_layout // FIXME(sil-serialize-all)
-  @usableFromInline
-  internal struct Index {
-    // Assumption: we rely on NSDictionary.getObjects when being
-    // repeatedly called on the same NSDictionary, returning items in the same
-    // order every time.
-    // Similarly, the same assumption holds for NSSet.allObjects.
-
-    /// A reference to the NSSet, which owns members in `allObjects`,
-    /// or `allKeys`, for NSSet and NSDictionary respectively.
-    @usableFromInline // FIXME(sil-serialize-all)
-    internal let base: _CocoaSet
-    // FIXME: swift-3-indexing-model: try to remove the cocoa reference, but
-    // make sure that we have a safety check for accessing `allKeys`.  Maybe
-    // move both into the dictionary/set itself.
-
-    /// An unowned array of keys.
-    @usableFromInline // FIXME(sil-serialize-all)
-    internal var allKeys: _HeapBuffer<Int, AnyObject>
-
-    /// Index into `allKeys`
-    @usableFromInline // FIXME(sil-serialize-all)
-    internal var currentKeyIndex: Int
-
-    @inlinable // FIXME(sil-serialize-all)
-    internal init(_ base: _CocoaSet, startIndex: ()) {
-      self.base = base
-      self.allKeys = _stdlib_NSSet_allObjects(base.object)
-      self.currentKeyIndex = 0
-    }
-
-    @inlinable // FIXME(sil-serialize-all)
-    internal init(_ base: _CocoaSet, endIndex: ()) {
-      self.base = base
-      self.allKeys = _stdlib_NSSet_allObjects(base.object)
-      self.currentKeyIndex = allKeys.value
-    }
-
-    @inlinable // FIXME(sil-serialize-all)
-    internal init(
-      _ base: _CocoaSet,
-      _ allKeys: _HeapBuffer<Int, AnyObject>,
-      _ currentKeyIndex: Int
-    ) {
-      self.base = base
-      self.allKeys = allKeys
-      self.currentKeyIndex = currentKeyIndex
-    }
-  }
-}
-
-extension _CocoaSet.Index: Equatable {
-  @inlinable
-  internal static func == (lhs: _CocoaSet.Index, rhs: _CocoaSet.Index) -> Bool {
-    _precondition(lhs.base.object === rhs.base.object,
-      "Comparing indexes from different sets")
-    return lhs.currentKeyIndex == rhs.currentKeyIndex
-  }
-}
-
-extension _CocoaSet.Index: Comparable {
-  @inlinable
-  internal static func < (lhs: _CocoaSet.Index, rhs: _CocoaSet.Index) -> Bool {
-    _precondition(lhs.base.object === rhs.base.object,
-      "Comparing indexes from different sets")
-    return lhs.currentKeyIndex < rhs.currentKeyIndex
-  }
-}
-#endif
-
 extension Set {
   /// The position of an element in a set.
   @_fixed_layout
@@ -3051,125 +1419,6 @@ extension Set.Index: Hashable {
   }
 }
 
-extension _NativeSet: Sequence {
-  @usableFromInline
-  @_fixed_layout
-  internal struct Iterator {
-    // The iterator is iterating over a frozen view of the collection state, so
-    // it keeps its own reference to the set.
-    @usableFromInline
-    internal let base: _NativeSet
-    @usableFromInline
-    internal var iterator: _HashTable.Iterator
-
-    @inlinable
-    init(_ base: _NativeSet) {
-      self.base = base
-      self.iterator = base.hashTable.makeIterator()
-    }
-  }
-
-  @inlinable
-  __consuming internal func makeIterator() -> Iterator {
-    return Iterator(self)
-  }
-}
-
-extension _NativeSet.Iterator: IteratorProtocol {
-  @inlinable
-  internal mutating func next() -> Element? {
-    guard let index = iterator.next() else { return nil }
-    return base.uncheckedElement(at: index)
-  }
-}
-
-#if _runtime(_ObjC)
-extension _CocoaSet: Sequence {
-  @usableFromInline
-  final internal class Iterator {
-    // Cocoa Set iterator has to be a class, otherwise we cannot
-    // guarantee that the fast enumeration struct is pinned to a certain memory
-    // location.
-
-    // This stored property should be stored at offset zero.  There's code below
-    // relying on this.
-    internal var _fastEnumerationState: _SwiftNSFastEnumerationState =
-      _makeSwiftNSFastEnumerationState()
-
-    // This stored property should be stored right after
-    // `_fastEnumerationState`.  There's code below relying on this.
-    internal var _fastEnumerationStackBuf = _CocoaFastEnumerationStackBuf()
-
-    internal let base: _CocoaSet
-
-    internal var _fastEnumerationStatePtr:
-      UnsafeMutablePointer<_SwiftNSFastEnumerationState> {
-      return _getUnsafePointerToStoredProperties(self).assumingMemoryBound(
-        to: _SwiftNSFastEnumerationState.self)
-    }
-
-    internal var _fastEnumerationStackBufPtr:
-      UnsafeMutablePointer<_CocoaFastEnumerationStackBuf> {
-      return UnsafeMutableRawPointer(_fastEnumerationStatePtr + 1)
-        .assumingMemoryBound(to: _CocoaFastEnumerationStackBuf.self)
-    }
-
-    // These members have to be word-sized integers, they cannot be limited to
-    // Int8 just because our storage holds 16 elements: fast enumeration is
-    // allowed to return inner pointers to the container, which can be much
-    // larger.
-    internal var itemIndex: Int = 0
-    internal var itemCount: Int = 0
-
-    internal init(_ base: _CocoaSet) {
-      self.base = base
-    }
-  }
-
-  @usableFromInline
-  __consuming internal func makeIterator() -> Iterator {
-    return Iterator(self)
-  }
-}
-
-extension _CocoaSet.Iterator: IteratorProtocol {
-  @usableFromInline
-  internal typealias Element = AnyObject
-
-  @usableFromInline
-  internal func next() -> Element? {
-    if itemIndex < 0 {
-      return nil
-    }
-    let base = self.base
-    if itemIndex == itemCount {
-      let stackBufCount = _fastEnumerationStackBuf.count
-      // We can't use `withUnsafeMutablePointer` here to get pointers to
-      // properties, because doing so might introduce a writeback storage, but
-      // fast enumeration relies on the pointer identity of the enumeration
-      // state struct.
-      itemCount = base.object.countByEnumerating(
-        with: _fastEnumerationStatePtr,
-        objects: UnsafeMutableRawPointer(_fastEnumerationStackBufPtr)
-          .assumingMemoryBound(to: AnyObject.self),
-        count: stackBufCount)
-      if itemCount == 0 {
-        itemIndex = -1
-        return nil
-      }
-      itemIndex = 0
-    }
-    let itemsPtrUP =
-    UnsafeMutableRawPointer(_fastEnumerationState.itemsPtr!)
-      .assumingMemoryBound(to: AnyObject.self)
-    let itemsPtr = _UnmanagedAnyObjectArray(itemsPtrUP)
-    let key: AnyObject = itemsPtr[itemIndex]
-    itemIndex += 1
-    return key
-  }
-}
-#endif
-
 extension Set {
   /// An iterator over the members of a `Set<Element>`.
   @_fixed_layout
@@ -3291,45 +1540,6 @@ extension Set: CustomReflectable {
   }
 }
 
-/// Initializes a `Set` from unique members.
-///
-/// Using a builder can be faster than inserting members into an empty
-/// `Set`.
-@_fixed_layout
-public // SPI(Foundation)
-struct _SetBuilder<Element: Hashable> {
-  @usableFromInline
-  internal var _target: _NativeSet<Element>
-  @usableFromInline
-  internal let _requestedCount: Int
-
-  @inlinable
-  public init(count: Int) {
-    _target = _NativeSet(capacity: count)
-    _requestedCount = count
-  }
-
-  @inlinable
-  public mutating func add(member: Element) {
-    _precondition(_target.count < _requestedCount,
-      "Can't add more members than promised")
-    _target.insertNew(member, isUnique: true)
-  }
-
-  @inlinable
-  public mutating func take() -> Set<Element> {
-    _precondition(_target.capacity > 0 || _requestedCount == 0,
-      "Cannot take the result twice")
-    _precondition(_target.count == _requestedCount,
-      "The number of members added does not match the promised count")
-
-    // Prevent taking the result twice.
-    var result = _NativeSet<Element>()
-    swap(&result, &_target)
-    return Set(_native: result)
-  }
-}
-
 extension Set {
   /// Removes and returns the first element of the set.
   ///
@@ -3369,46 +1579,6 @@ extension Set {
     _sanityCheck(self.capacity >= minimumCapacity)
   }
 }
-
-//===--- Bridging ---------------------------------------------------------===//
-
-#if _runtime(_ObjC)
-extension Set {
-  @inlinable
-  public func _bridgeToObjectiveCImpl() -> _NSSetCore {
-    switch _variant {
-    case .native(let nativeSet):
-      return nativeSet.bridged()
-    case .cocoa(let cocoaSet):
-      return cocoaSet.object
-    }
-  }
-
-  /// Returns the native Dictionary hidden inside this NSDictionary;
-  /// returns nil otherwise.
-  public static func _bridgeFromObjectiveCAdoptingNativeStorageOf(
-    _ s: AnyObject
-  ) -> Set<Element>? {
-
-    // Try all three NSSet impls that we currently provide.
-
-    if let deferred = s as? _SwiftDeferredNSSet<Element> {
-      return Set(_native: deferred.native)
-    }
-
-    if let nativeStorage = s as? _SetStorage<Element> {
-      return Set(_native: _NativeSet(nativeStorage))
-    }
-
-    if s === _RawSetStorage.empty {
-      return Set()
-    }
-
-    // FIXME: what if `s` is native storage, but for different key/value type?
-    return nil
-  }
-}
-#endif
 
 public typealias SetIndex<Element: Hashable> = Set<Element>.Index
 public typealias SetIterator<Element: Hashable> = Set<Element>.Iterator
